refactor: NotificationService, alarmReceiver 위치 수정
#186
