+++ app/src/main/java/com/boosters/promise/receiver/alarm/di/AlarmModule.kt
@@ -1,7 +1,7 @@
-package com.boosters.promise.ui.notification.di
+package com.boosters.promise.receiver.alarm.di
 
 import android.content.Context
-import com.boosters.promise.ui.notification.AlarmDirector
+import com.boosters.promise.receiver.alarm.AlarmDirector
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
