+++ app/src/main/java/com/boosters/promise/service/notification/BootService.kt
@@ -1,10 +1,11 @@
-package com.boosters.promise.ui.notification
+package com.boosters.promise.service.notification
 
 import android.content.Intent
 import androidx.lifecycle.LifecycleService
 import androidx.lifecycle.lifecycleScope
 import com.boosters.promise.data.promise.PromiseRepository
 import com.boosters.promise.data.user.UserRepository
+import com.boosters.promise.receiver.alarm.AlarmDirector
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
