+++ app/src/main/java/com/boosters/promise/ui/detail/PromiseDetailViewModel.kt
@@ -17,8 +17,8 @@ import com.boosters.promise.ui.detail.model.MemberMarkerInfo
 import com.boosters.promise.ui.detail.model.MemberUiModel
 import com.boosters.promise.ui.detail.model.PromiseFailUiState
 import com.boosters.promise.ui.detail.model.PromiseUploadUiState
-import com.boosters.promise.ui.notification.AlarmDirector
-import com.boosters.promise.ui.notification.NotificationService
+import com.boosters.promise.receiver.alarm.AlarmDirector
+import com.boosters.promise.service.notification.NotificationService
 import com.boosters.promise.util.DateUtil
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
